+++ conftest.py
@@ -49,7 +49,8 @@ def clean_up_test_files():
         'dt.cc2',
         'dt.ct',
         'dt.ct2',
-        'phase.dat'
+        'phase.dat',
+        '2016p008194.xml'
     ]
 
     yield